feat(middleware)!: new method to delete duplicates notifications by types has been created in the middleware

+++ src/backend/notifications_service/notifications_app/middleware.py
@@ -6,7 +6,7 @@ from django.http import JsonResponse
 from django.conf import settings
 from .models import Notification
 from django.utils import timezone
-from datetime import timedelta
+from datetime import timedelta, datetime
  
 User = get_user_model()
 # Middleware for jwt authentication
@@ -52,13 +52,42 @@ class JWTAuthMiddleware(MiddlewareMixin):
  
 class NotificationMiddleware(MiddlewareMixin):
     def process_request(self, request):
+        self.remove_duplicate_notifications(request)
         try:
             user_notifications = list(Notification.objects.filter(receiver=request.user, is_read=True))
-            for notification in user_notifications:
-                if notification.is_read_at and notification.is_read_at < timezone.now() - timedelta(minutes=1):  #replace by timedelta(days=????)
+            for notification in user_notifications:  
+                if notification.type == 'friend-request-accepted' and notification.is_read_at and notification.is_read_at < timezone.now() - timedelta(minutes=1):  #replace by timedelta(days=????)
                     notification.delete()
         except Exception as e:
             pass
         response = self.get_response(request)
         return response
     
+    
+    def remove_duplicate_notifications(self, request):
+        try:
+            user_notifications = Notification.objects.filter(receiver=request.user)
+            senders = set(user_notification.sender for user_notification in user_notifications)
+            for sender in senders:
+                notifications_by_sender = user_notifications.filter(sender=sender)
+                self.remove_duplicate_notifications_by_sender(notifications_by_sender)         
+        except Exception as e: 
+            pass
+        
+        
+    def remove_duplicate_notifications_by_sender(self, notifications_by_sender):
+        friend_request_accepted = list(notifications_by_sender.filter(type='friend-request-accepted'))
+        friend_request_pending = list(notifications_by_sender.filter(type='friend-request-pending'))
+        private_match_invitation = list(notifications_by_sender.filter(type='private-match-invitation'))
+        tournament_invitation = list(notifications_by_sender.filter(type='tournament-invitation'))
+        
+        self.delete_notifications(friend_request_accepted[:-1])
+        self.delete_notifications(friend_request_pending[:-1])
+        self.delete_notifications(private_match_invitation[:-1])
+        self.delete_notifications(tournament_invitation[:-1])                   
+        
+    
+    def delete_notifications(self, notifications_to_delete):
+        for notification in notifications_to_delete:
+            notification.delete()
+
